[minor]added mongodbsettings.py in app.py
