+++ python/src/mas/cli/install/app.py
@@ -568,6 +568,34 @@ class InstallApp(BaseApp, InstallSettingsMixin, InstallSummarizerMixin, ConfigGe
                 self.promptForString("IBM Cloud API Key", "ibmcloud_apikey", isPassword=True)
                 self.promptForString("IBM Cloud Resource Group", "cos_resourcegroup")
 
+    def configMongoDb(self) -> None:
+
+             self.printDescription([
+                "The installer can setup mongoce in your OpenShift cluster (available only for amd64 )or you may choose to configure MAS to use an existing mongodb"
+             ])
+             self.printH1("Configure MongoDb")
+             self.promptForString("Install namespace", "mongodb_namespace", default="mongoce")
+
+
+             if self.yesOrNo("CConfigure MongoDb in your OpenShift cluster"):
+                self.setParam("mongo_action_system", "install")
+             else:
+                mongodb_namespace = 'mongoce'
+                self.setParam("mongo_action_system", "byo")
+                self.selectLocalConfigDir()
+
+                instanceId = self.getParam('mas_instance_id')
+                # Check if a configuration already exists before creating a new one
+                mongoCfgFile = path.join(self.localConfigDir, f"mongo-{mongodb_namespace}.yaml")
+                print_formatted_text(f"Searching for system database configuration file in {mongoCfgFile} ...")
+                if path.exists(mongoCfgFile):
+                    if self.yesOrNo(f"System database configuration file 'mongo-{instanceId}-system.yam' already exists.  Do you want to generate a new one"):
+                        self.generateMongoCfg(instanceId=instanceId, destination=mongoCfgFile)
+                else:
+                    print_formatted_text(f"Expected file ({mongoCfgFile}) was not found, generating a valid system database configuration file now ...")
+                    self.generateMongoCfg(instanceId=instanceId,destination=mongoCfgFile)
+
+
     def interactiveMode(self) -> None:
         # Interactive mode
         self.interactiveMode = True
@@ -604,8 +632,7 @@ class InstallApp(BaseApp, InstallSettingsMixin, InstallSummarizerMixin, ConfigGe
         #if not self.preview:
         self.configMongoDb()  # Will only do anything if IoT or Manage have been selected for install
         self.configDb2()
-        self.configKafka()  # Will only do anything if IoT has been selected for install
-
+        self.configKafka()
         self.configGrafana()
         self.configTurbonomic()
